fix(store): Fix integer overflow in get_into/batch_get_into for values > 4GB (#920)
* Fix integer overflow in get_into and batch_get_into for values > 4GB

Problem:
- get_into() and batch_get_into() returned 'int' (32-bit signed)
- This caused integer overflow when handling values larger than ~2GB
- Values > 4GB would wrap around and return incorrect results

Solution:
- Changed return type from 'int' to 'int64_t' for both methods
- Updated batch_get_into to return std::vector<int64_t>
- Updated test assertions to use int64_t

Impact:
- Can now handle values up to ~9.2 exabytes (2^63 - 1)
- Backward compatible for values < 2GB
- Internal implementation already used int64_t, so this aligns public API
- Python bindings automatically handle int64_t -> Python int conversion

Files changed:
- mooncake-store/include/pybind_client.h
- mooncake-store/src/pybind_client.cpp
- mooncake-store/tests/pybind_client_test.cpp

* docs: Update Doxygen comment for batch_get_into return type

Update the documentation comment to accurately reflect that batch_get_into
returns a vector of 64-bit integers, not just 'integers'. This clarifies
that the return type can handle values larger than 2GB/4GB.

+++ mooncake-store/include/pybind_client.h
@@ -110,7 +110,7 @@ class PyClient {
      * @note The buffer address must be previously registered with
      * register_buffer() for zero-copy operations
      */
-    int get_into(const std::string &key, void *buffer, size_t size);
+    int64_t get_into(const std::string &key, void *buffer, size_t size);
 
     /**
      * @brief Get object data directly into pre-allocated buffers for multiple
@@ -118,14 +118,14 @@ class PyClient {
      * @param keys Vector of keys of the objects to get
      * @param buffers Vector of pointers to the pre-allocated buffers
      * @param sizes Vector of sizes of the buffers
-     * @return Vector of integers, where each element is the number of bytes
-     * read on success, or a negative value on error
+     * @return Vector of 64-bit integers, where each element is the number of
+     * bytes read on success, or a negative value on error
      * @note The buffer addresses must be previously registered with
      * register_buffer() for zero-copy operations
      */
-    std::vector<int> batch_get_into(const std::vector<std::string> &keys,
-                                    const std::vector<void *> &buffers,
-                                    const std::vector<size_t> &sizes);
+    std::vector<int64_t> batch_get_into(const std::vector<std::string> &keys,
+                                        const std::vector<void *> &buffers,
+                                        const std::vector<size_t> &sizes);
 
     /**
      * @brief Put object data directly from a pre-allocated buffer
